rename: collector repository 의존성 변경
- core 모듈의 repository -> collector 모듈의 자체 repository interface로 변경
