+++ collector/src/main/java/kernel360/ckt/collector/CollectorApplication.java
@@ -3,11 +3,9 @@ package kernel360.ckt.collector;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ComponentScan;
 
 
 @EntityScan(basePackages = "kernel360.ckt.core")
-@ComponentScan(basePackages = {"kernel360.ckt.core", "kernel360.ckt.collector"})
 @SpringBootApplication
 public class CollectorApplication {
     public static void main(String[] args) {
