+++ collector/src/main/java/kernel360/ckt/collector/application/repository/RentalRepository.java
@@ -1,0 +1,12 @@
+package kernel360.ckt.collector.application.repository;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+import kernel360.ckt.core.domain.entity.RentalEntity;
+import kernel360.ckt.core.domain.enums.RentalStatus;
+
+public interface RentalRepository {
+
+    Optional<RentalEntity> findFirstByVehicleIdAndStatusAndPickupAt(Long vehicleId, RentalStatus status, LocalDateTime pickupAt);
+
+}
