+++ collector/src/main/java/kernel360/ckt/collector/application/repository/VehicleTraceLogRepository.java
@@ -1,0 +1,7 @@
+package kernel360.ckt.collector.application.repository;
+
+import kernel360.ckt.core.domain.entity.VehicleTraceLogEntity;
+
+public interface VehicleTraceLogRepository {
+    VehicleTraceLogEntity save(VehicleTraceLogEntity vehicleTraceLog);
+}
