+++ collector/src/main/java/kernel360/ckt/collector/infra/repository/jpa/RouteJpaRepository.java
@@ -1,7 +1,7 @@
 package kernel360.ckt.collector.infra.repository.jpa;
 
+import kernel360.ckt.collector.application.repository.RouteRepository;
 import kernel360.ckt.core.domain.entity.RouteEntity;
-import kernel360.ckt.core.repository.RouteRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RouteJpaRepository extends JpaRepository<RouteEntity, Long>, RouteRepository {
