+++ collector/src/main/java/kernel360/ckt/collector/infra/repository/jpa/VehicleTraceLogJpaRepository.java
@@ -1,8 +1,9 @@
 package kernel360.ckt.collector.infra.repository.jpa;
 
+import kernel360.ckt.collector.application.repository.VehicleTraceLogRepository;
 import kernel360.ckt.core.domain.entity.VehicleTraceLogEntity;
-import kernel360.ckt.core.repository.VehicleTraceLogRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleTraceLogJpaRepository extends JpaRepository<VehicleTraceLogEntity, Long>, VehicleTraceLogRepository {
+public interface VehicleTraceLogJpaRepository extends JpaRepository<VehicleTraceLogEntity, Long>,
+    VehicleTraceLogRepository {
 }
